back-end/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-sendgrid-v5==0.9.1
dnspython==2.1.0
future==0.18.2
gunicorn==20.1.0
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.6.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.2.0
yapf==0.31.0
